MR errors fix
