+++ src/cmds/cpp/ctest_cmd.rs
@@ -1,26 +1,8 @@
-//! Filters ctest output — passing tests collapsed, failures shown verbatim.
+//! Filters ctest output — keep GoogleTest failures and final summaries.
 
 use crate::core::runner::{self, RunOptions};
 use crate::core::utils::resolved_command;
 use anyhow::Result;
-use lazy_static::lazy_static;
-use regex::Regex;
-
-lazy_static! {
-    // "Test #N: name ........... Passed   0.42 sec"
-    static ref PASS_RE: Regex = Regex::new(r"^\s*Test\s+#\d+:\s+\S+.*\bPassed\b").unwrap();
-    // "Start N: TestName"
-    static ref START_RE: Regex = Regex::new(r"^\s*Start\s+\d+:\s+\S+").unwrap();
-    // "100% tests passed, 0 tests failed out of 15"
-    static ref SUMMARY_RE: Regex =
-        Regex::new(r"^\s*\d+%\s+tests passed,\s+\d+\s+tests? failed out of \d+").unwrap();
-    // "Total Test time (real) =   0.42 sec"
-    static ref TIME_RE: Regex = Regex::new(r"^\s*Total Test time").unwrap();
-    // "The following tests FAILED:"
-    static ref FAILED_HEADER_RE: Regex = Regex::new(r"^\s*The following tests FAILED:").unwrap();
-    // Failed test entry: "    1 - TestName (Failed)"
-    static ref FAILED_ENTRY_RE: Regex = Regex::new(r"^\s+\d+\s+-\s+\S+\s+\(.*\)").unwrap();
-}
 
 pub fn run(args: &[String], verbose: u8) -> Result<i32> {
     let mut cmd = resolved_command("ctest");
@@ -34,212 +16,346 @@ pub fn run(args: &[String], verbose: u8) -> Result<i32> {
         cmd,
         "ctest",
         &args.join(" "),
-        filter_output,
+        filter_ctest_output,
         RunOptions::with_tee("ctest"),
     )
 }
 
-fn filter_output(raw: &str) -> String {
-    let mut summary: Option<String> = None;
-    let mut total_time: Option<String> = None;
-    let mut failures: Vec<String> = Vec::new();
-    let mut failure_capture: Vec<String> = Vec::new();
-    let mut in_failed_block = false;
-    let mut in_output_block = false;
-    let mut current_output: Vec<String> = Vec::new();
-    let mut passed = 0usize;
-    let mut total_tests = 0usize;
+pub(crate) fn filter_ctest_output(output: &str) -> String {
+    let mut out: Vec<String> = Vec::new();
+    let mut current_test: Option<String> = None;
+    let mut emitting_failure_block = false;
+    let mut emitted_test_header_for_block = false;
 
-    for line in raw.lines() {
-        if PASS_RE.is_match(line) {
-            passed += 1;
-            total_tests += 1;
-            in_output_block = false;
-            if !current_output.is_empty() {
-                current_output.clear();
-            }
-            continue;
-        }
-        if START_RE.is_match(line) {
-            total_tests = total_tests.max(extract_test_number(line));
-            continue;
-        }
-        if TIME_RE.is_match(line) {
-            total_time = Some(line.trim().to_string());
-            continue;
-        }
-        if let Some(s) = parse_summary(line) {
-            summary = Some(s);
-            continue;
-        }
-        if FAILED_HEADER_RE.is_match(line) {
-            in_failed_block = true;
-            failure_capture.push(line.to_string());
-            continue;
-        }
-        if in_failed_block && (FAILED_ENTRY_RE.is_match(line) || line.trim().is_empty()) {
-            failure_capture.push(line.to_string());
-            if line.trim().is_empty() {
-                in_failed_block = false;
-            }
-            continue;
-        }
-        if line.contains("Failed") && line.contains("Test #") {
-            failures.push(line.to_string());
-            in_output_block = true;
-            current_output.clear();
-            continue;
-        }
-        if in_output_block {
-            // Captured failed-test output ends at the next "Test #" line or summary block
-            if PASS_RE.is_match(line) || line.trim().starts_with("Total Test time") {
-                in_output_block = false;
-                failure_capture.append(&mut current_output);
-            } else {
-                current_output.push(line.to_string());
-            }
-        }
-    }
+    for line in output.lines() {
+        let trimmed = line.trim_end();
+        let t = trimmed.trim();
+        let m = normalize_ctest_verbose_prefix(t);
 
-    if !current_output.is_empty() {
-        failure_capture.extend(current_output);
-    }
-
-    let mut out = String::new();
-    if failures.is_empty() && failure_capture.is_empty() {
-        let time = total_time
-            .as_deref()
-            .and_then(extract_seconds)
-            .unwrap_or_default();
-        if let Some(s) = summary {
-            out.push_str(&format!("ctest: {}", s));
-            if !time.is_empty() {
-                out.push_str(&format!("  {}", time));
+        if t.is_empty() {
+            if emitting_failure_block {
+                out.push(String::new());
             }
-        } else if total_tests > 0 {
-            out.push_str(&format!("ctest: {}/{} passed", passed, total_tests));
-            if !time.is_empty() {
-                out.push_str(&format!("  {}", time));
-            }
-        } else {
-            out.push_str("ctest: ok");
+            continue;
         }
-        return out;
+
+        if let Some(name) = parse_gtest_run(m) {
+            current_test = Some(name);
+            emitting_failure_block = false;
+            emitted_test_header_for_block = false;
+            continue;
+        }
+
+        if is_gtest_ok(m) {
+            emitting_failure_block = false;
+            emitted_test_header_for_block = false;
+            current_test = None;
+            continue;
+        }
+
+        if is_gtest_failure_header(m) {
+            if !emitted_test_header_for_block {
+                if let Some(name) = current_test.as_deref() {
+                    out.push(name.to_string());
+                }
+                emitted_test_header_for_block = true;
+            }
+            emitting_failure_block = true;
+            out.push(trimmed.to_string());
+            continue;
+        }
+
+        if is_gtest_failed_summary_line(m) {
+            emitting_failure_block = false;
+            emitted_test_header_for_block = false;
+            out.push(trimmed.to_string());
+            continue;
+        }
+
+        if is_ctest_summary_line(m) {
+            emitting_failure_block = false;
+            emitted_test_header_for_block = false;
+            out.push(trimmed.to_string());
+            continue;
+        }
+
+        if is_important_non_gtest_line(m) {
+            out.push(trimmed.to_string());
+            continue;
+        }
+
+        if emitting_failure_block {
+            if is_gtest_failed_test_line(m) {
+                out.push(trimmed.to_string());
+                emitting_failure_block = false;
+                emitted_test_header_for_block = false;
+                current_test = None;
+                continue;
+            }
+
+            if is_noisy_separator(m) {
+                continue;
+            }
+
+            out.push(trimmed.to_string());
+        }
     }
 
-    for f in &failures {
-        out.push_str(f.trim_end());
-        out.push('\n');
-    }
-    if !failure_capture.is_empty() {
-        out.push('\n');
-        for line in &failure_capture {
-            out.push_str(line.trim_end());
-            out.push('\n');
-        }
-    }
-    if let Some(s) = summary {
-        out.push('\n');
-        out.push_str(&s);
-    }
-    out
+    compact_lines(out).join("\n")
 }
 
-fn extract_test_number(line: &str) -> usize {
-    line.split_whitespace()
-        .find_map(|t| t.trim_end_matches(':').parse::<usize>().ok())
-        .unwrap_or(0)
+fn normalize_ctest_verbose_prefix(line: &str) -> &str {
+    let s = line.trim_start();
+    let mut i = 0;
+    for (idx, ch) in s.char_indices() {
+        if ch.is_ascii_digit() {
+            i = idx + ch.len_utf8();
+            continue;
+        }
+        i = idx;
+        break;
+    }
+    if i == 0 {
+        return line;
+    }
+    let rest = &s[i..];
+    if !rest.starts_with(':') {
+        return line;
+    }
+    let rest = &rest[1..];
+    let rest = rest.strip_prefix(' ').unwrap_or(rest);
+    if rest.is_empty() {
+        line
+    } else {
+        rest
+    }
 }
 
-fn parse_summary(line: &str) -> Option<String> {
-    if !SUMMARY_RE.is_match(line) {
+fn parse_gtest_run(line: &str) -> Option<String> {
+    if !line.starts_with("[ RUN") {
         return None;
     }
-    Some(line.trim().to_string())
+    let idx = line.find("]")?;
+    let rest = line[idx + 1..].trim();
+    if rest.is_empty() {
+        None
+    } else {
+        Some(rest.to_string())
+    }
 }
 
-fn extract_seconds(line: &str) -> Option<String> {
-    line.split('=').nth(1).map(|s| s.trim().to_string())
+fn is_gtest_ok(line: &str) -> bool {
+    line.starts_with("[       OK ]")
+}
+
+fn is_gtest_failure_header(line: &str) -> bool {
+    line.ends_with(": Failure")
+}
+
+fn is_gtest_failed_test_line(line: &str) -> bool {
+    line.starts_with("[  FAILED  ]") && line.contains('(')
+}
+
+fn is_gtest_failed_summary_line(line: &str) -> bool {
+    if line.starts_with("[  FAILED  ]") {
+        return true;
+    }
+    if line.starts_with("[  PASSED  ]") && line.contains("tests") {
+        return true;
+    }
+    false
+}
+
+fn is_ctest_summary_line(line: &str) -> bool {
+    let l = line.to_lowercase();
+    l.contains("tests failed")
+        || l.contains("% tests passed")
+        || l.contains("the following tests failed")
+        || l.contains("errors while running ctest")
+        || l.contains("ctest error")
+}
+
+fn is_important_non_gtest_line(line: &str) -> bool {
+    let l = line.to_lowercase();
+    l.contains("addresssanitizer")
+        || l.contains("runtime error")
+        || l.contains("undefined behavior")
+        || l.contains("segmentation fault")
+        || l.contains("abort")
+        || l.starts_with("error:")
+        || l.contains("fatal error")
+        || l.contains("cmake error")
+        || l.contains("clang: error")
+        || l.contains("gcc: error")
+        || l.contains("ld: error")
+}
+
+fn is_noisy_separator(line: &str) -> bool {
+    line.chars().all(|c| c == '=' || c == '-' || c == '─' || c == '═')
+}
+
+fn compact_lines(mut lines: Vec<String>) -> Vec<String> {
+    while lines.first().is_some_and(|l| l.trim().is_empty()) {
+        lines.remove(0);
+    }
+    while lines.last().is_some_and(|l| l.trim().is_empty()) {
+        lines.pop();
+    }
+
+    let mut out: Vec<String> = Vec::with_capacity(lines.len());
+    let mut last_blank = false;
+    for l in lines.drain(..) {
+        let blank = l.trim().is_empty();
+        if blank {
+            if last_blank {
+                continue;
+            }
+            last_blank = true;
+            out.push(String::new());
+        } else {
+            last_blank = false;
+            out.push(l);
+        }
+    }
+    out
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
 
-    fn count_tokens(s: &str) -> usize {
-        s.split_whitespace().count()
+    #[test]
+    fn gtest_all_passed_compacts() {
+        let input = r#"
+[==========] Running 2 tests from 1 test suite.
+[ RUN      ] Foo.Pass
+[       OK ] Foo.Pass (1 ms)
+[ RUN      ] Foo.Pass2
+[       OK ] Foo.Pass2 (1 ms)
+[==========] 2 tests from 1 test suite ran. (2 ms total)
+[  PASSED  ] 2 tests.
+"#;
+        let out = filter_ctest_output(input);
+        assert!(!out.contains("Foo.Pass"));
+        assert!(!out.contains("[ RUN"));
+        assert!(out.contains("[  PASSED  ] 2 tests."));
     }
 
     #[test]
-    fn test_all_pass() {
-        let raw = "Test project /tmp/build\n\
-            Start 1: test_one\n\
-            1/3 Test #1: test_one ......................   Passed    0.01 sec\n\
-            Start 2: test_two\n\
-            2/3 Test #2: test_two ......................   Passed    0.01 sec\n\
-            Start 3: test_three\n\
-            3/3 Test #3: test_three ....................   Passed    0.01 sec\n\
-            \n\
-            100% tests passed, 0 tests failed out of 3\n\
-            \n\
-            Total Test time (real) =   0.42 sec\n";
-        let out = filter_output(raw);
-        assert!(out.starts_with("ctest:"));
-        assert!(out.contains("100%") || out.contains("3/3"));
-        assert!(!out.contains("Start 1:"));
+    fn gtest_one_failed_keeps_block_and_summary() {
+        let input = r#"
+[==========] Running 3 tests from 1 test suite.
+[ RUN      ] Foo.Pass
+[       OK ] Foo.Pass (1 ms)
+[ RUN      ] Foo.Fail
+/path/foo_test.cc:42: Failure
+Expected equality of these values:
+  actual
+  expected
+[  FAILED  ] Foo.Fail (0 ms)
+[ RUN      ] Foo.Pass2
+[       OK ] Foo.Pass2 (1 ms)
+[==========] 3 tests from 1 test suite ran.
+[  PASSED  ] 2 tests.
+[  FAILED  ] 1 test, listed below:
+[  FAILED  ] Foo.Fail
+"#;
+        let out = filter_ctest_output(input);
+        assert!(out.contains("Foo.Fail"));
+        assert!(out.contains("/path/foo_test.cc:42: Failure"));
+        assert!(out.contains("Expected equality of these values:"));
+        assert!(out.contains("[  FAILED  ] 1 test, listed below:"));
+        assert!(out.contains("[  FAILED  ] Foo.Fail"));
+        assert!(!out.contains("Foo.Pass"));
+        assert!(!out.contains("Foo.Pass2"));
     }
 
     #[test]
-    fn test_failure_keeps_summary() {
-        let raw = "Start 1: test_a\n\
-            1/2 Test #1: test_a ......................   Passed    0.01 sec\n\
-            Start 2: test_b\n\
-            2/2 Test #2: test_b ......................***Failed    0.05 sec\n\
-                assertion failed: expected 4, got 5\n\
-            \n\
-            50% tests passed, 1 tests failed out of 2\n\
-            \n\
-            The following tests FAILED:\n\
-            \t  2 - test_b (Failed)\n\
-            \n\
-            Errors while running CTest\n";
-        let out = filter_output(raw);
-        assert!(out.contains("test_b"));
-        assert!(out.contains("Failed"));
-        assert!(out.contains("50% tests passed"));
-        assert!(!out.contains("Start 1:"));
+    fn gtest_multiple_failed_keeps_both_blocks() {
+        let input = r#"
+[ RUN      ] A.Fail
+/p/a.cc:1: Failure
+boom
+[  FAILED  ] A.Fail (0 ms)
+[ RUN      ] B.Fail
+/p/b.cc:2: Failure
+kaboom
+[  FAILED  ] B.Fail (0 ms)
+[  FAILED  ] 2 tests, listed below:
+[  FAILED  ] A.Fail
+[  FAILED  ] B.Fail
+"#;
+        let out = filter_ctest_output(input);
+        assert!(out.contains("A.Fail"));
+        assert!(out.contains("/p/a.cc:1: Failure"));
+        assert!(out.contains("B.Fail"));
+        assert!(out.contains("/p/b.cc:2: Failure"));
+        assert!(out.contains("[  FAILED  ] A.Fail"));
+        assert!(out.contains("[  FAILED  ] B.Fail"));
     }
 
     #[test]
-    fn test_fixture_success() {
-        let raw = include_str!("../../../tests/fixtures/cpp/ctest_success.txt");
-        let out = filter_output(raw);
-        assert!(out.starts_with("ctest:"));
-        assert!(!out.contains("Start  1:"));
+    fn gtest_parameterized_names_survive() {
+        let input = r#"
+[ RUN      ] FooTest/0.DoesThing
+/p/t.cc:3: Failure
+nope
+[  FAILED  ] FooTest/0.DoesThing (0 ms)
+[ RUN      ] FooSuite/BarTest.DoesThing/1
+/p/u.cc:4: Failure
+nope2
+[  FAILED  ] FooSuite/BarTest.DoesThing/1 (0 ms)
+[  FAILED  ] 2 tests, listed below:
+[  FAILED  ] FooTest/0.DoesThing
+[  FAILED  ] FooSuite/BarTest.DoesThing/1
+"#;
+        let out = filter_ctest_output(input);
+        assert!(out.contains("FooTest/0.DoesThing"));
+        assert!(out.contains("FooSuite/BarTest.DoesThing/1"));
     }
 
     #[test]
-    fn test_fixture_failure() {
-        let raw = include_str!("../../../tests/fixtures/cpp/ctest_failure.txt");
-        let out = filter_output(raw);
-        assert!(out.contains("test_parser_failure"));
-        assert!(out.contains("test_runtime_failure"));
-        assert!(out.contains("60% tests passed"));
+    fn keeps_non_gtest_important_lines() {
+        let input = r#"
+[ RUN      ] Foo.Fail
+/p/t.cc:3: Failure
+nope
+AddressSanitizer: heap-use-after-free
+[  FAILED  ] Foo.Fail (0 ms)
+"#;
+        let out = filter_ctest_output(input);
+        assert!(out.contains("AddressSanitizer: heap-use-after-free"));
+        assert!(out.contains("/p/t.cc:3: Failure"));
     }
 
     #[test]
-    fn test_savings() {
-        let mut raw = String::new();
-        for i in 1..=20 {
-            raw.push_str(&format!("Start {}: test_{}\n", i, i));
-            raw.push_str(&format!(
-                "{}/20 Test #{}: test_{} ........................   Passed    0.01 sec\n",
-                i, i, i
-            ));
-        }
-        raw.push_str("100% tests passed, 0 tests failed out of 20\n");
-        let out = filter_output(&raw);
-        let savings = 100.0 - (count_tokens(&out) as f64 / count_tokens(&raw) as f64 * 100.0);
-        assert!(savings >= 60.0, "expected >=60%, got {:.1}%", savings);
+    fn ctest_verbose_prefix_gtest_failure_survives_and_pass_noise_drops() {
+        let input = r#"
+1: [ RUN      ] Foo.Pass
+1: [       OK ] Foo.Pass (1 ms)
+1: [ RUN      ] Foo.Fail
+1: /path/foo_test.cc:42: Failure
+1: Expected equality of these values:
+1:   actual
+1:   expected
+1: [  FAILED  ] Foo.Fail (0 ms)
+1: [ RUN      ] Foo.Pass2
+1: [       OK ] Foo.Pass2 (1 ms)
+1: [  FAILED  ] 1 test, listed below:
+1: [  FAILED  ] Foo.Fail
+"#;
+        let out = filter_ctest_output(input);
+        assert!(out.contains("Foo.Fail"));
+        assert!(out.contains("1: /path/foo_test.cc:42: Failure"));
+        assert!(out.contains("1: Expected equality of these values:"));
+        assert!(out.contains("1: [  FAILED  ] Foo.Fail"));
+        assert!(!out.contains("Foo.Pass"));
+        assert!(!out.contains("Foo.Pass2"));
+    }
+
+    #[test]
+    fn ctest_verbose_prefix_normalizer_requires_colon() {
+        assert_eq!(normalize_ctest_verbose_prefix("123 tests failed"), "123 tests failed");
+        assert_eq!(normalize_ctest_verbose_prefix("404 error happened"), "404 error happened");
     }
 }
